Add the Google style guide and formatting options
This adds the Google styleguide as a build time check (it will fail if you violate the guide) and adds the same style guide to the format document command.

Some useful commands to know (find in the Ctrl+shift+p menu):
Organize Imports
Format Document

F2 on a variable to rename it and rename all references to it.